+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Form validation @robgietema
 - Notification messages @robgietema
 
+### Changes
+- Updated to new history api @robgietema
+
 ## 0.3.0 (2017-04-29)
 ### Added
 - Personal information @robgietema